feat: Add orders API and integrate frontend order flow
Introduce server-side endpoints and model support for paginated user orders and reviews, and wire up the frontend order UI and review flow. Backend: add Api\V1\OrderController, OrderRepository#getPaginatedUserOrders, Order model filterByStatus scope, rename OrderItem->review relation, update Product/Product resources for numeric casting, and remove the IOrderRepository contract. Frontend: add orders API client and queries (list + review mutation), integrate account routes to expose Orders page, refactor account sidebar/nav labels, remove account overview and wishlist pages/components, implement paginated OrderList/OrderCard/OrderItemCard to consume new API shape (OrderResponse, OrderItem), add review modal wiring to submit reviews and invalidate queries, and minor UI/typing adjustments. These changes enable listing user orders with pagination and submitting item reviews.

+++ backend/app/Http/Resources/V1/ProductDetailResource.php
@@ -32,7 +32,7 @@ final class ProductDetailResource extends JsonResource
                 'max' => $this->product_items_max_price,
             ],
             'rating' => [
-                'average' => round($this->avg_rating ?? 0, 1),
+                'average' => round((float) ($this->avg_rating ?? 0), 1),
                 'count' => $this->reviews_count ?? 0,
             ],
             'total_stock' => (int) $this->product_items_sum_qty_in_stock,
